+++ rapyer/init.py
@@ -1,10 +1,15 @@
 import logging
 
 import redis.asyncio as redis_async
-from rapyer.base import REDIS_MODELS
-from rapyer.scripts import register_scripts
 from redis import ResponseError
 from redis.asyncio.client import Redis
+
+from rapyer.base import REDIS_MODELS
+from rapyer.scripts import register_scripts
+
+
+def is_fakeredis(client) -> bool:
+    return "fakeredis" in type(client).__module__
 
 
 async def init_rapyer(
@@ -24,12 +29,14 @@ async def init_rapyer(
     if isinstance(redis, str):
         redis = redis_async.from_url(redis, decode_responses=True, max_connections=20)
 
+    is_fake_redis = is_fakeredis(redis)
     if redis is not None:
-        await register_scripts(redis)
+        await register_scripts(redis, is_fake_redis)
 
     for model in REDIS_MODELS:
         if redis is not None:
             model.Meta.redis = redis
+            model.Meta.is_fake_redis = is_fake_redis
         if ttl is not None:
             model.Meta.ttl = ttl
         if prefer_normal_json_dump is not None: