+++ rapyer/types/dct.py
@@ -131,7 +131,12 @@ class RedisDict(dict[str, T], GenericRedisType, Generic[T]):
 
     def __setitem__(self, key, value):
         if self.pipeline:
-            self.pipeline.json().set(self.key, self.json_field_path(key), value)
+            serialized = self._adapter.dump_python(
+                {key: value}, mode="json", context={REDIS_DUMP_FLAG_NAME: True}
+            )
+            self.pipeline.json().set(
+                self.key, self.json_field_path(key), serialized[key]
+            )
         new_val = self.validate_dict({key: value})[key]
         super().__setitem__(key, new_val)